Moving iou_with_jaccard and jaccard from mean_average_precision.py to detection/utils.py
